refactor(controllers): promote duplicated internal/ packages to shared core/controllers/internal/ (CC1, #1095)
Slice CC1 of EPIC-0 (#1095) — Coordinator-led consolidation. The 5 Group C
controllers (slices C1-C5: organization, environment, blueprint, application,
useraccess) all merged with their own per-controller go.mod + per-controller
internal/ tree. This PR canonicalizes the shared layout per
`02-implementer-canon.md` §1+§2:

  * One go.mod at core/controllers/go.mod (Path A — single shared module)
  * Shared helpers under core/controllers/internal/:
      - semver/    (was: blueprint/internal/semver + application/internal/semver,
                    now exposes blueprint's IsValidRange + app's IsExact, with
                    the union of both test corpora)
      - placement/ (was: application/internal/placement; promoted per seam map)
      - render/    (was: application/internal/render; promoted per seam map)
      - labels/    (was: useraccess/internal/labels; promoted per seam map —
                    Manara-style scope matcher, owner-of-record C5)

Module-discipline decision (Path A vs Path B): Path A. The 5 controllers'
go.mod files use the same controller-runtime v0.19.0, k8s.io/* @ 0.31.x,
sigs.k8s.io/yaml v1.4.0, etc. The only drift was organization-controller
on k8s.io/api 0.31.0 vs the others on 0.31.1 — a trivial bump.
Independent dep-version pinning would only be valuable if a controller
needed a hostile dep the others shouldn't pull; nothing in the current
tree is hostile.

Containerfiles + workflows updated:
  * 5 Containerfiles now COPY core/controllers/{go.mod,go.sum,internal/}
    plus the per-controller tree from a repo-root build context.
  * 4 per-controller workflows (application/environment/organization/
    useraccess; blueprint-controller has no dedicated workflow yet) now
    trigger on core/controllers/{<name>/**, internal/**, go.mod, go.sum}
    and run go vet + go test scoped to their own tree + shared internal.
  * useraccess workflow context flipped from core/controllers/useraccess
    to . (repo root) so the Containerfile can reach the shared go.mod.

Subpackages NOT promoted in this PR (compromise — flagged for follow-up):
  * gitea/ — 4 of 5 controllers each ship a Gitea HTTP client. The APIs
    DIVERGE (organization has Org+Repo CRUD with Repo struct return values;
    application/blueprint/environment have File CRUD with Org-not-found
    sentinel). A SUPERSET package would require renaming methods (e.g.
    EnsureRepo collides on signature) which crosses the brief's "no API
    redesign" line. CC2 follow-up slice should design the unified surface
    before promoting.
  * validate/ — application's package validates Application.spec.parameters
    against a JSON Schema (santhosh-tekuri lib); blueprint's validates
    Blueprint CR business rules (semver-backed). Same dir name, completely
    different functions — not actually duplicates.
  * gitops/ — environment's renders Flux GitRepository for an Environment;
    organization's renders HelmRelease+Namespace for an Org. Same dir name,
    different inputs and outputs.

Test-coverage delta: pre-consolidation 134 root-level tests (sum across
5 modules); post-consolidation 133 tests. Net delta -1: blueprint and
application each had their own TestIsValidRange in their semver pkg; the
shared semver pkg's TestIsValidRange now exercises the union of both
controllers' valid+invalid input corpora — coverage strictly improved
even though one redundant test name disappeared.

Verified locally: go build + go vet + `go test -count=1 -race ./...`
all clean; all 5 controller binaries (cmd/) link successfully.
