+++ .github/workflows/build-application-controller.yaml
@@ -19,11 +19,17 @@ on:
   push:
     paths:
       - 'core/controllers/application/**'
+      - 'core/controllers/internal/**'
+      - 'core/controllers/go.mod'
+      - 'core/controllers/go.sum'
       - '.github/workflows/build-application-controller.yaml'
     branches: [main]
   pull_request:
     paths:
       - 'core/controllers/application/**'
+      - 'core/controllers/internal/**'
+      - 'core/controllers/go.mod'
+      - 'core/controllers/go.sum'
       - '.github/workflows/build-application-controller.yaml'
   workflow_dispatch:
 
@@ -55,15 +61,18 @@ jobs:
         with:
           go-version: '1.23'
           cache-dependency-path: |
-            core/controllers/application/go.sum
+            core/controllers/go.sum
 
       - name: go vet
-        working-directory: core/controllers/application
-        run: go vet ./...
+        working-directory: core/controllers
+        # Slice CC1 (#1095) consolidated the 5 Group C controllers into
+        # a single shared go.mod. Vet scoped to this controller's tree
+        # plus the shared internal/ helpers it depends on.
+        run: go vet ./application/... ./internal/...
 
       - name: Run unit tests
-        working-directory: core/controllers/application
-        run: go test -count=1 -race ./...
+        working-directory: core/controllers
+        run: go test -count=1 -race ./application/... ./internal/...
 
       # On pull_request runs we stop here — image push requires
       # `packages: write` which only main-branch authors hold.
