+++ .github/workflows/build-organization-controller.yaml
@@ -19,11 +19,17 @@ on:
   push:
     paths:
       - 'core/controllers/organization/**'
+      - 'core/controllers/internal/**'
+      - 'core/controllers/go.mod'
+      - 'core/controllers/go.sum'
       - '.github/workflows/build-organization-controller.yaml'
     branches: [main]
   pull_request:
     paths:
       - 'core/controllers/organization/**'
+      - 'core/controllers/internal/**'
+      - 'core/controllers/go.mod'
+      - 'core/controllers/go.sum'
       - '.github/workflows/build-organization-controller.yaml'
   workflow_dispatch:
 
@@ -54,15 +60,18 @@ jobs:
         with:
           go-version: '1.23'
           cache-dependency-path: |
-            core/controllers/organization/go.sum
+            core/controllers/go.sum
 
       - name: Vet
-        working-directory: core/controllers/organization
-        run: go vet ./...
+        working-directory: core/controllers
+        # Slice CC1 (#1095) consolidated the 5 Group C controllers into
+        # a single shared go.mod. Vet scoped to this controller's tree
+        # plus the shared internal/ helpers it depends on.
+        run: go vet ./organization/... ./internal/...
 
       - name: Test
-        working-directory: core/controllers/organization
-        run: go test -count=1 -race ./...
+        working-directory: core/controllers
+        run: go test -count=1 -race ./organization/... ./internal/...
 
       - name: Login to GHCR
         if: github.event_name != 'pull_request'
