+++ core/controllers/application/internal/controller/application_controller.go
@@ -59,9 +59,9 @@ import (
 	"k8s.io/apimachinery/pkg/watch"
 	"k8s.io/client-go/dynamic"
 
-	"github.com/openova-io/openova/core/controllers/application/internal/placement"
-	"github.com/openova-io/openova/core/controllers/application/internal/render"
-	"github.com/openova-io/openova/core/controllers/application/internal/semver"
+	"github.com/openova-io/openova/core/controllers/internal/placement"
+	"github.com/openova-io/openova/core/controllers/internal/render"
+	"github.com/openova-io/openova/core/controllers/internal/semver"
 	"github.com/openova-io/openova/core/controllers/application/internal/validate"
 )
