+++ core/controllers/internal/labels/scope.go
