+++ core/controllers/internal/render/manifests.go
