+++ core/controllers/internal/render/manifests_test.go
