+++ core/controllers/internal/semver/semver_test.go
@@ -6,7 +6,8 @@ func TestIsValidRange(t *testing.T) {
 	t.Parallel()
 
 	// Cases gathered from a sweep of every existing
-	// platform/*/blueprint.yaml `upgrades.from[]` entry.
+	// platform/*/blueprint.yaml `upgrades.from[]` entry. Slice CC1
+	// merged the C3 (blueprint) and C4 (application) test corpora.
 	valid := []string{
 		"0.x",
 		"1.x",
@@ -18,6 +19,7 @@ func TestIsValidRange(t *testing.T) {
 		"~1.4",
 		"1.0.0",
 		"1.2.3",
+		"1.2.3-beta.1",
 		"1.0.0-rc.1",
 		"1.0.0-beta-2",
 		"1.0.0+build.5",
@@ -43,6 +45,7 @@ func TestIsValidRange(t *testing.T) {
 		"abc",
 		"1.2.3.4",
 		"1..2",
+		"1.0..0",
 		"^",
 		"~",
 		">=",
@@ -57,6 +60,30 @@ func TestIsValidRange(t *testing.T) {
 	for _, s := range invalid {
 		if err := IsValidRange(s); err == nil {
 			t.Errorf("IsValidRange(%q) = nil, want error", s)
+		}
+	}
+}
+
+func TestIsExact(t *testing.T) {
+	t.Parallel()
+
+	cases := []struct {
+		in   string
+		want bool
+	}{
+		{"1.0.0", true},
+		{"1.2.3-beta.1", true},
+		{"1.0", false},
+		{"^1.0.0", false},
+		{"1.0.x", false},
+		{"1.x", false},
+		{"", false},
+		{"abc", false},
+	}
+	for _, tc := range cases {
+		got := IsExact(tc.in)
+		if got != tc.want {
+			t.Errorf("IsExact(%q) = %v, want %v", tc.in, got, tc.want)
 		}
 	}
 }
