+++ core/controllers/useraccess/internal/controller/spec_test.go
@@ -5,7 +5,7 @@ import (
 
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
-	"github.com/openova-io/openova/core/controllers/useraccess/internal/labels"
+	"github.com/openova-io/openova/core/controllers/internal/labels"
 )
 
 func mkUA(spec map[string]any) *unstructured.Unstructured {
